# Backend/app/api/v1/routers/tests.py
from typing import List
from fastapi import APIRouter, Depends, HTTPException, status
from sqlalchemy.ext.asyncio import AsyncSession

from app.api.v1.access import (
    get_manageable_test,
    get_user_level_context,
    get_visible_test,
    is_unlocked_test,
)
from app.api.deps import get_db
from app.cache.redis_cache import (
    NS_MATERIALS,
    NS_TESTS,
    NS_TEST_CONTENT,
    NS_TEST_SUMMARY,
    TEST_CONTENT_TTL,
    TESTS_LIST_TTL,
    TEST_DETAIL_TTL,
    TEST_SUMMARY_TTL,
    bump_cache_namespace,
    cache_key_test_content,
    cache_key_test_detail,
    cache_key_test_list,
    cache_key_test_summary,
    get_cache_namespace_version,
    get,
    set,
)
from app.core.security import get_current_user, require_roles
from app.models.user import User
from app.schemas.grading import AttemptScoreUpdate
from app.schemas.question import QuestionRead
from app.schemas.test_ import TestCreate, TestRead, TestUpdate
from app.schemas.analytics import TestSummary
from app.schemas.test_attempt import TestAttemptRead
from app.schemas.test_content import TestContentRead
from app.repositories import analytics_repo, test_repo, test_attempt_repo
from app.repositories import question_repo
from app.services import test_service
from app.services.test_runtime import AttemptPolicyError, finalize_attempt_if_expired, resolve_attempt_for_user

router = APIRouter()


@router.get("/", response_model=List[TestRead], status_code=status.HTTP_200_OK)
async def list_tests(
    published_only: bool = True,
    limit: int = 100,
    db: AsyncSession = Depends(get_db),
    current_user: User = Depends(get_current_user),
):
    if not published_only and current_user.role not in {"teacher", "admin"}:
        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Insufficient permissions")

    _, level_id = await get_user_level_context(db, current_user)
    if published_only:
        version = await get_cache_namespace_version(NS_TESTS)
        cache_key = cache_key_test_list(
            published_only=published_only,
            limit=limit,
            level_id=level_id,
            version=version,
        )
        cached = await get(cache_key)
        if cached is not None:
            return cached

    author_id = current_user.id if (not published_only and current_user.role == "teacher") else None
    items = await test_repo.list_tests(db, published_only=published_only, limit=limit, author_id=author_id)
    if published_only and current_user.role not in {"teacher", "admin"}:
        items = [item for item in items if await is_unlocked_test(db, current_user, item)]
    if published_only:
        payload = [TestRead.model_validate(item).model_dump(mode="json") for item in items]
        await set(cache_key, payload, ttl=TESTS_LIST_TTL)
    return items


@router.get("/{test_id}", response_model=TestRead, status_code=status.HTTP_200_OK)
async def get_test(
    test_id: int,
    db: AsyncSession = Depends(get_db),
    current_user: User = Depends(get_current_user),
):
    level_id = 0
    if current_user.role not in {"teacher", "admin"}:
        _, level_id = await get_user_level_context(db, current_user)
        version = await get_cache_namespace_version(NS_TESTS)
        cache_key = cache_key_test_detail(test_id, level_id=level_id, version=version)
        cached = await get(cache_key)
        if cached is not None:
            return cached

    t = await get_visible_test(db, test_id, current_user)
    if t.published:
        payload = TestRead.model_validate(t).model_dump(mode="json")
        version = await get_cache_namespace_version(NS_TESTS)
        await set(cache_key_test_detail(test_id, level_id=level_id, version=version), payload, ttl=TEST_DETAIL_TTL)
    return t


@router.post("/", response_model=TestRead, status_code=status.HTTP_201_CREATED)
async def create_test(
    payload: TestCreate,
    db: AsyncSession = Depends(get_db),
    current_user: User = Depends(require_roles("teacher", "admin")),
):
    test = await test_service.create_test(db, payload, current_user)
    await bump_cache_namespace(NS_TESTS, NS_TEST_CONTENT, NS_TEST_SUMMARY, NS_MATERIALS)
    return test


@router.patch("/{test_id}", response_model=TestRead, status_code=status.HTTP_200_OK)
async def update_test(
    test_id: int,
    payload: TestUpdate,
    db: AsyncSession = Depends(get_db),
    current_user: User = Depends(require_roles("teacher", "admin")),
):
    test = await test_service.update_test(db, test_id, payload, current_user)
    await bump_cache_namespace(NS_TESTS, NS_TEST_CONTENT, NS_TEST_SUMMARY, NS_MATERIALS)
    return test


@router.post("/{test_id}/publish", response_model=TestRead, status_code=status.HTTP_200_OK)
async def publish_test(
    test_id: int,
    db: AsyncSession = Depends(get_db),
    current_user: User = Depends(require_roles("teacher", "admin")),
):
    await get_manageable_test(db, test_id, current_user)
    test = await test_repo.update_test(db, test_id, published=True)
    if test is None:
        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Test not found")
    await bump_cache_namespace(NS_TESTS, NS_TEST_CONTENT, NS_TEST_SUMMARY)
    return test


@router.post("/{test_id}/hide", response_model=TestRead, status_code=status.HTTP_200_OK)
async def hide_test(
    test_id: int,
    db: AsyncSession = Depends(get_db),
    current_user: User = Depends(require_roles("teacher", "admin")),
):
    await get_manageable_test(db, test_id, current_user)
    test = await test_repo.update_test(db, test_id, published=False)
    if test is None:
        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Test not found")
    await bump_cache_namespace(NS_TESTS, NS_TEST_CONTENT, NS_TEST_SUMMARY)
    return test


@router.delete("/{test_id}", status_code=status.HTTP_204_NO_CONTENT)
async def delete_test(
    test_id: int,
    db: AsyncSession = Depends(get_db),
    current_user: User = Depends(require_roles("teacher", "admin")),
):
    await get_manageable_test(db, test_id, current_user)
    deleted = await test_repo.delete_test(db, test_id)
    if not deleted:
        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Test not found")
    await bump_cache_namespace(NS_TESTS, NS_TEST_CONTENT, NS_TEST_SUMMARY, NS_MATERIALS)
    return {}


@router.get("/{test_id}/summary", response_model=TestSummary, status_code=status.HTTP_200_OK)
async def test_summary(
    test_id: int,
    db: AsyncSession = Depends(get_db),
    current_user: User = Depends(get_current_user),
):
    is_teacher = current_user.role in {"teacher", "admin"}
    summary_version = await get_cache_namespace_version(NS_TEST_SUMMARY)
    cache_key = cache_key_test_summary(test_id, version=summary_version)
    test = await test_service.get_test_or_summary_access(db, test_id, current_user)
    if not is_teacher:
        cached = await get(cache_key)
        if cached is not None:
            return cached

    summary = await test_repo.get_test_summary(db, test_id)
    if test.published:
        await set(cache_key, summary, ttl=TEST_SUMMARY_TTL)
    return summary


@router.get("/{test_id}/content", response_model=TestContentRead, status_code=status.HTTP_200_OK)
async def get_test_content(
    test_id: int,
    db: AsyncSession = Depends(get_db),
    current_user: User = Depends(get_current_user),
):
    is_teacher = current_user.role in {"teacher", "admin"}
    _, level_id = await get_user_level_context(db, current_user)
    content_version = await get_cache_namespace_version(NS_TEST_CONTENT)
    cache_key = cache_key_test_content(test_id, level_id=level_id, version=content_version)
    if not is_teacher:
        cached = await get(cache_key)
        if cached is not None:
            return cached

    test = await get_visible_test(db, test_id, current_user)

    questions = await question_repo.list_questions_for_test(db, test_id=test_id, limit=500, offset=0)
    payload = {
        "test": TestRead.model_validate(test).model_dump(mode="json"),
        "questions": [QuestionRead.model_validate(question).model_dump(mode="json") for question in questions],
    }
    if test.published:
        await set(cache_key, payload, ttl=TEST_CONTENT_TTL)
    return payload


@router.post("/{test_id}/attempts/start", response_model=TestAttemptRead, status_code=status.HTTP_201_CREATED)
async def start_test_attempt(
    test_id: int,
    db: AsyncSession = Depends(get_db),
    current_user: User = Depends(get_current_user),
):
    test = await get_visible_test(db, test_id, current_user)
    try:
        return await resolve_attempt_for_user(db, test, current_user.id)
    except AttemptPolicyError as exc:
        raise HTTPException(status_code=status.HTTP_409_CONFLICT, detail=str(exc))


@router.get("/{test_id}/attempts/me", response_model=List[TestAttemptRead], status_code=status.HTTP_200_OK)
async def list_my_test_attempts(
    test_id: int,
    db: AsyncSession = Depends(get_db),
    current_user: User = Depends(get_current_user),
):
    await get_visible_test(db, test_id, current_user)
    return await test_attempt_repo.list_attempts_for_user(db, current_user.id, test_id=test_id)


@router.post("/attempts/{attempt_id}/complete", response_model=TestAttemptRead, status_code=status.HTTP_200_OK)
async def complete_test_attempt(
    attempt_id: int,
    db: AsyncSession = Depends(get_db),
    current_user: User = Depends(get_current_user),
):
    attempt = await test_attempt_repo.get_attempt(db, attempt_id)
    if attempt is None:
        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Attempt not found")
    if attempt.user_id != current_user.id and current_user.role not in {"teacher", "admin"}:
        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Insufficient permissions")
    test = await test_service.get_test_or_summary_access(db, attempt.test_id, current_user)
    if current_user.role == "teacher":
        await get_manageable_test(db, attempt.test_id, current_user)
    if attempt.status == "completed":
        return attempt
    _, _ = await finalize_attempt_if_expired(db, test, attempt)
    if attempt.status == "completed":
        return attempt
    completed_attempt = await test_attempt_repo.complete_attempt(db, attempt)
    await analytics_repo.register_completed_attempt(db, completed_attempt.user_id)
    return completed_attempt


@router.patch("/attempts/{attempt_id}/score", response_model=TestAttemptRead, status_code=status.HTTP_200_OK)
async def override_attempt_score(
    attempt_id: int,
    payload: AttemptScoreUpdate,
    db: AsyncSession = Depends(get_db),
    current_user: User = Depends(require_roles("teacher", "admin")),
):
    attempt = await test_attempt_repo.get_attempt(db, attempt_id)
    if attempt is None:
        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Attempt not found")
    if current_user.role == "teacher":
        await get_manageable_test(db, attempt.test_id, current_user)
    if attempt.status != "completed":
        raise HTTPException(status_code=status.HTTP_409_CONFLICT, detail="Attempt must be completed before final grading")

    if attempt.max_score is None:
        attempt = await test_attempt_repo.refresh_attempt_scores(db, attempt)
    max_score = float(attempt.max_score or 0.0)
    if payload.score < 0 or payload.score > max_score:
        raise HTTPException(
            status_code=status.HTTP_400_BAD_REQUEST,
            detail=f"Score must be between 0 and {max_score}",
        )

    return await test_attempt_repo.set_manual_score(db, attempt, payload.score)