# Backend/docs/api-contract.yaml
openapi: 3.0.3
info:
  title: HSE C# Gamification Backend API
  version: 1.0.0
  description: Frontend-oriented API contract for the current backend.
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TestRead:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: string, nullable: true }
        time_limit_minutes: { type: integer, nullable: true }
        max_score: { type: integer, nullable: true }
        published: { type: boolean }
        published_at: { type: string, format: date-time, nullable: true }
        material_id: { type: integer, nullable: true }
        material_ids:
          type: array
          items: { type: integer }
        deadline: { type: string, format: date-time, nullable: true }
        author_id: { type: integer, nullable: true }
        required_level_id: { type: integer, nullable: true }
    TestCreate:
      type: object
      required: [title]
      properties:
        title: { type: string }
        description: { type: string, nullable: true }
        time_limit_minutes: { type: integer, nullable: true }
        max_score: { type: integer, nullable: true }
        published: { type: boolean }
        material_id: { type: integer, nullable: true }
        material_ids:
          type: array
          items: { type: integer }
        deadline: { type: string, format: date-time, nullable: true }
        required_level_id: { type: integer, nullable: true }
...
    AttemptScoreUpdate:
      type: object
      required: [score]
      properties:
        score: { type: number, format: float }
        comment: { type: string, nullable: true }
    TestAttemptRead:
      type: object
      properties:
        id: { type: integer }
        user_id: { type: integer }
        test_id: { type: integer }
        status: { type: string }
        score: { type: number, format: float, nullable: true }
        manual_score: { type: number, format: float, nullable: true }
        max_score: { type: number, format: float, nullable: true }
        time_spent_seconds: { type: integer, nullable: true }
        started_at: { type: string, format: date-time }
        submitted_at: { type: string, format: date-time, nullable: true }
        completed_at: { type: string, format: date-time, nullable: true }
...
paths:
...
  /api/v1/tests:
    get:
      summary: List tests, published only by default
      description: For regular users published tests are additionally filtered by required level.
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/TestRead' }
    post:
      summary: Create test
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TestCreate' }
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TestRead' }
...
  /api/v1/tests/{test_id}/attempts/start:
    post:
      summary: Start attempt or return active one
      description: One active attempt per test. Re-taking completed test is forbidden. Locked tests are inaccessible for regular users until their level is high enough.
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TestAttemptRead' }
        '409': { description: Deadline passed, time limit exceeded, or test already completed }
...
  /api/v1/tests/attempts/{attempt_id}/score:
    patch:
      summary: Manual override for completed attempt score
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AttemptScoreUpdate' }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TestAttemptRead' }
        '409': { description: Attempt must be completed }
...